# perf(chunk_range): compute bitmap once