Compute available-missing intersection once, then pop mins iteratively. Reduces O(n) clones/Ands to O(1) on hot path, boosting batch request efficiency for large queues (e.g., 16-32 chunks/peer).

**Before**

 chunks_test.go
```
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent
cpu: Apple M1 Pro
BenchmarkChunksPopBatch-10         20325         59230 ns/op       57896 B/op       2960 allocs/op
PASS
ok      github.com/james-lawrence/torrent   2.075s
```


After

chunks_test.go

```
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent
cpu: Apple M1 Pro
BenchmarkChunksPopBatch-10        169858          7146 ns/op        3831 B/op        108 allocs/op
PASS
ok      github.com/james-lawrence/torrent   2.563s
```

(And confirming existing test works below) 

```
Running tool: /opt/homebrew/opt/go@1.23/bin/go test -timeout 30s -run ^TestChunksPop$ github.com/james-lawrence/torrent

ok      github.com/james-lawrence/torrent   0.262s
```