@MaanasSathaye 2025-08-10
@james-lawrence here's your peekn. you were assigning the same cidx to all in the slice